refactor: refactor components & class names
