+++ src/components/NumberBoard/number-board.module.scss
@@ -1,13 +1,13 @@
 @import '../../styles/variables.scss';
 
-.numberBoardRoot {
+.container {
 	display: flex;
 	flex-direction: row;
 	justify-content: center;
 	flex-wrap: wrap;
 }
 
-.numberInput {
+.button {
 	flex-basis: 20%;
 	max-width: 3rem;
 	height: 3rem;
@@ -21,24 +21,25 @@
 	color: $foreground_nom;
 	border-radius: 50%;
 	border: 2px solid rgba(white, 0.12);
+	transition: all 0.2s ease;
 
-	&.numberInputFilled {
+	&.filled {
 		color: $primary_color;
 	}
 
 	&:hover {
-		color: $foreground_high;
-		border-color: $primary_color;
+		color: $primary_color;
+		border-color: rgb($primary_color, 0.2);
 		background-color: rgb($primary_color, 0.2);
 	}
 
-	&:focus {
+	&:focus-visible {
 		color: $foreground_high;
 		border-color: $primary_color;
 		background-color: rgba($primary_color, 0.4);
 	}
 
-	&.numberInputActive {
+	&.active {
 		color: $on_primary;
 		border-color: $primary_color;
 		background-color: $primary_color;
