--- src/components/Sudoku/index.ts
@@ -1,1 +1,0 @@
-export * from './Sudoku'
