+++ src/components/SudokuBoard/components/SudokuCell/sudoku-cell.module.scss
@@ -1,4 +1,4 @@
-@import '../../styles/variables.scss';
+@import '../../../../styles/variables.scss';
 
 .sudokuCell {
 	width: 1.5rem;
