+++ src/components/SudokuBoard/components/SudokuSquare/sudoku-square.module.scss
@@ -1,4 +1,4 @@
-@import '../../styles/variables.scss';
+@import '../../../../styles/variables.scss';
 
 .sudokuSquare {
 	display: grid;
