+++ src/components/SudokuBoard/components/SudokuSquare/types.d.ts
@@ -1,5 +1,5 @@
 import {Dispatch, SetStateAction} from 'react'
-import {Cell} from '../../types/sudoku'
+import {Cell} from '../../../../types/sudoku'
 
 export type SudokuSquareProps = {
 	squareRow: number
