+++ src/components/SudokuBoard/components/index.ts
@@ -1,0 +1,2 @@
+export * from './SudokuCell'
+export * from './SudokuSquare'
