+++ src/components/SudokuRenderer/SudokuRenderer.tsx
@@ -1,0 +1,10 @@
+import {SudokuWrapper} from '..'
+import {SudokuContextProvider} from '../../contexts'
+
+export function SudokuRenderer() {
+	return (
+		<SudokuContextProvider>
+			<SudokuWrapper />
+		</SudokuContextProvider>
+	)
+}
