--- src/components/SudokuRoot/index.ts
@@ -1,1 +1,0 @@
-export * from './SudokuRoot'
