+++ src/components/SudokuWrapper/sudoku-wrapper.module.scss
@@ -1,0 +1,5 @@
+.wrapper {
+	display: flex;
+	flex-direction: column;
+	align-items: center;
+}
