+++ src/components/index.ts
@@ -1,6 +1,5 @@
-export * from './Sudoku'
-export * from './SudokuCell'
-export * from './SudokuSquare'
+export * from './SudokuBoard'
 export * from './NumberBoard'
-export * from './SudokuRoot'
+export * from './SudokuWrapper'
+export * from './SudokuRenderer'
 export * from './WinGreet'
